requirements.txt
Flask==2.3.2
pymongo==3.12.0
scikit-learn==1.3.0
numpy==1.24.0
flask-cors==5.0.0
python-dotenv==1.0.0
dnspython==2.2.1
google-generativeai
PyPDF2
